Refacto GeojsonTiler architecture
* Split Geojson class in GeojsonPolygon and GeojsonLine classes
* Read files in GeojsonTiler.py script
* Split line buffering from Geojson class
